A try to reduce code dup. for attribute validators
