+++ spec/fortnox/api/validators/attributes/examples_for_validate.rb
@@ -1,15 +1,25 @@
-shared_examples_for '.validate' do |attribute|
+shared_examples_for '.validate' do |attribute, model_class, validator_class|
+  using_test_classes do
+    class Model < Fortnox::API::Model::Base
+    end
+    Model.send( :include, model_class )
+
+    class Validator < Fortnox::API::Validator::Base
+    end
+    Validator.send( :include, validator_class )
+  end
+
   describe '.validate' do
-    subject{ validator_class.new }
+    subject{ Validator.new }
 
     context "with invalid #{attribute}" do
-      let( :invalid_model ){ model_class.new( attribute => invalid_attribute ) }
+      let( :invalid_model ){ Model.new( attribute => invalid_attribute ) }
 
       specify{ expect(subject.validate( invalid_model )).to eql( false ) }
     end
 
     context "with valid #{attribute}" do
-      let( :valid_model ){ model_class.new( attribute => valid_attribute  ) }
+      let( :valid_model ){ Model.new( attribute => valid_attribute  ) }
 
       specify{ expect( subject.validate( valid_model )).to eql( true ) }
     end
